requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.0.4
colorama==0.4.6
exceptiongroup==1.3.0
flask==2.2.1
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.0.1
jinja2==3.0.3
keras==2.10.0
keras-preprocessing==1.1.2
libclang==18.1.1
markdown==3.9
markupsafe==3.0.3
numpy==1.23.5
oauthlib==3.3.1
opt-einsum==3.4.0
packaging==25.0
pillow==9.3.0
pluggy==1.6.0
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
pytest==8.3.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tomli==2.2.1
typing-extensions==4.15.0
urllib3==2.5.0
werkzeug==2.2.3
wrapt==1.17.3
selenium
webdriver-manager==3.8.6
Locust
